+++ Visuals/Spectogram/YAxis.h
@@ -1,0 +1,21 @@
+
+/** $VER: YAxis.h (2024.03.29) P. Stuer - Implements the Y axis of a spectogram. **/
+
+#pragma once
+
+#include <CppCoreCheck/Warnings.h>
+
+#pragma warning(disable: 4100 4625 4626 4710 4711 5045 ALL_CPPCORECHECK_WARNINGS)
+
+#include <SDKDDKVer.h>
+#include <WinSock2.h>
+#include <Windows.h>
+
+/// <summary>
+/// Implements the Y axis of a spectogram.
+/// </summary>
+#pragma warning(disable: 4820)
+class SpectogramYAxis : public Element
+{
+public:
+};
